feat: add description for VIP-220
